pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ocr-app"
version = "0.1.0"
description = "Сервис суммаризации документов: извлечение текста из PDF/изображений и создание саммари через языковые модели"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "PolyForm Noncommercial License 1.0.0"}
authors = [
    {name = "NaumovAO", email = "workpostNA@yandex.ru"}
]
maintainers = [
    {name = "NaumovAO", email = "workpostNA@yandex.ru"}
]
keywords = ["ocr", "summarization", "llm", "document-processing", "russian-nlp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: Other/Proprietary License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing",
    "Topic :: Utilities"
]

dependencies = [
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "pillow>=12.1.0",
    "pdfplumber>=0.11.9",
    "pytesseract>=0.3.13",
    "tenacity>=9.1.2",
    "pydantic>=2.12.5,<3.0.0",
    "pydantic-settings>=2.12.0,<3.0.0",
    "pyyaml>=6.0.3"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "mypy>=1.10.0",
    "pylint>=3.0.0",
    "types-requests",
    "types-PyYAML"
]
test = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0"
]

[project.scripts]
ocr-app = "ocr_app.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["ocr_app*"]
exclude = ["logs*", "source_data*", "output*", "test_*"]
namespaces = false

[tool.setuptools.package-data]
ocr_app = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | logs
  | output
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip = ["logs", "output", "source_data", ".venv", "venv"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--tb=short",
    "--doctest-modules",
    "--cov=ocr_app",
    "--cov-report=term-missing"
]